Hide cases if they haven't been synced into the API
We're currently having an issue on production where the my-cases API endpoint
is failing, because tenancies which haven't been synced are being requested.

This should log a warning and skip the tenancy, so the user can see all the valid
data we have without incomplete data.

Need to fix the sync to avoid this in future.