Add empty use cases to split letter functionality
This commit is intended as the first step of moving the 'save
a pdf to the cloud' functionality from 'letters#send_letter' to
'letters#create'.

As saving/loading from the cache will be redundant once this
move is complete, the caching behaviour is also worth pulling out
(so it can be easily removed).

Notably, this commit includes explanatory comments, as the
authors are getting their heads around the complex behaviour
of these components.