Move new use case spec to simpler folder
It may be correct to copy the existing use cases structure, or it
may not. To delay this decision, this commit (and commits working in
the same area) are placing new use cases in the simplest place -
the top-level use cases folder.

This commit has no opinion on where they could/should be moved later.