Refactor 'letter one' logic, add more tests
Refactoring this identified tests that we were missing - one for when
there's been no communication ever (which was already implemented but
not tested), and one for when the communication is old but the
arrears amount / other conditions aren't met (which is both tested and
implemented in this commit).

This commit also removes the upper boundary for arrears for letter
one, as otherwise something with a high amount of arrears but no
previous communication will never fall into any of these letter
actions.

Many of the conditions extracted into guard clauses in this commit
could be generic for many of these different classifications, but this
is for a later commit.

Likewise, some nosp-flow tests were removed as part of this commit as
they look like they 'should' be letter ones based on our new
understanding.