Fix 'old'-style tests around classifications
The work on 7c828f was done against the 'matrix'-style tests around
classifications, with a run of the full test suite being missed.

There's some preference for one test style over the other in our team,
but I think both test suites still have value as we're making lots of
changes in this area and learning a bunch as we go, so I'd rather keep
both working for now.