(Small) refactor of send_nosp method
We touched this method as part of a set of changes around pulling out
the 'warning letter' from the arrears process, and noticed some cheap
improvements. Here they are!

This commit includes a few extra scenarios in the specs, written to
help guide the implementation. I'm conscious of the proliferation of
these scenarios, but felt including them was better overall. In later
commits, I would like to cut down these examples to only the relevant
data, which should make these files less 'heavy' overall.