Remove superfluous SMS classification logic
If there's been no communication, then it'll never be the case
that a last communication fell outside the range being specified
in the removed line.

It's likely that the rules want to be more sophisticated than "don't
send an SMS if we've ever contacted them before", but this commit is
intended as a refactor that makes this gap more obvious, not a change
in behaviour.