Remove some default case information from stub
Most of these defaults aren't necessary for the specs to pass, but
they do sometimes cause newly written specs to give false positives.

This tripped me up a few times, and I don't want it to surprise the
next poor person who comes here!

I wasn't able to remove some of the defaults as some tests do rely on
them. Some of these specs are around unused functionality, but I did
not want to let the scope of this particular change grow.