Only get master accounts from Universal Housing
Some accounts in the `tenagree` table are sub-accounts of others, but
we don't want to retrieve these (and we certainly don't want to do
things like send them letters!).

Accounts that we do want to retrieve have a value of 'M' in
tenagree.agr_type. There is a lookup table to give this value more
semantic meaning, but joining onto this table probably gives us little
long-term benefit (as we're using the database as the integration
layer anyway).

Notably, this gateway is what we use for our 'sync', which pulls
across all the cases we show in our various worktrays in the service.

This commit also includes a small bit of cleanup on the query, which
passes the tests and shouldn't affect the raw SQL generated, it's just
a little easier to read.