Balance is in arrears when over the weekly gross rent
E.g. If I get charged my rent on Monday and I get paid on Friday, and I 
pay it then. I am not in arrears as I have paid my rent in the current 
week.

However, if I am a penny short. I am now in arrears on Monday by a 
penny.